+++ resources/views/cp/pages/index.blade.php
@@ -21,13 +21,15 @@
 			<div class="{{ $mainSmCols }} {{ $mainMdCols }} {{ $mainLgCols }} main">
 				@include('cp._partials.message')
 				@include('cp._partials.pageTitle')
-				<div class="row">
-					<div class="col">
-						<ul class="list-unstyled list-inline buttons">
-							<li class="list-inline-item"><a href="/cp/pages/create" title="Add Page" class="btn btn-outline-success"><i class="icon fa fa-plus" aria-hidden="true"></i>Add Page</a></li>
-						</ul>
+				@if ($currentUser->hasPermission('create_pages'))
+					<div class="row">
+						<div class="col">
+							<ul class="list-unstyled list-inline buttons">
+								<li class="list-inline-item"><a href="/cp/pages/create" title="Add Page" class="btn btn-outline-success"><i class="icon fa fa-plus" aria-hidden="true"></i>Add Page</a></li>
+							</ul>
+						</div>
 					</div>
-				</div>
+				@endif
 				<div class="content padding bg-white">	
 					<table id="datatable" class="table table-striped table-bordered table-hover" cellspacing="0" border="0" cellpadding="0" width="100%">
 						<thead>
@@ -51,8 +53,10 @@
 									<td class="actions dropdown text-center" id="submenu">
 										<a href="javascript:void(0);" title="Page Actions" class="dropdown-toggle" id="pageActions" data-toggle="dropdown"><i class="fa fa-ellipsis-v" aria-hidden="true"></i></a>
 										<ul class="actions dropdown-menu dropdown-menu-right">
-											<li class="dropdown-item gf-info"><a href="/cp/pages/{{ $page->id }}/edit" title="Edit Page"><i class="icon fa fa-pencil" aria-hidden="true"></i>Edit Page</a></li>
-											@if ($page->id > 1)
+											@if ($currentUser->hasPermission('edit_pages'))
+												<li class="dropdown-item gf-info"><a href="/cp/pages/{{ $page->id }}/edit" title="Edit Page"><i class="icon fa fa-pencil" aria-hidden="true"></i>Edit Page</a></li>
+											@endif
+											@if ($currentUser->hasPermission('delete_pages') && $page->id > 1)
 												<li class="dropdown-item gf-danger"><a href="/cp/pages/{{ $page->id }}/delete" title="Delete Page"><i class="icon fa fa-trash" aria-hidden="true"></i>Delete Page</a></li>
 											@endif
 										</ul>
